requirements.txt
nltk==3.2.3
numpy==1.12.1
Flask==0.12.2
scikit_learn==0.20.2
requests==2.18.4
gunicorn==19.9.0
more-itertools==6.0.0
spacyspacy>=2.0.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz